SQLControls will now take in the exception and parse out the check constraint.
Edit form adjusted to work with check constraint values and also check
to make sure that all fields are filled prior to submission.

If a field is empty that is an easy face-value check while the SQL commands
will return the constraint that was not met for the column that has the error.